Rename `default.py` to `recommenders.py`
